cambio de tipo de variable

+++ src/main/java/com/nttdata/transfers/service/Calculate.java
@@ -3,6 +3,6 @@ package com.nttdata.transfers.service;
 @FunctionalInterface
 public interface Calculate {
 
-    Float calcular(Float monto, Float saldo);
+    Double calcular(Double monto, Double saldo);
 
 }
